requirements.txt
Django
djangorestframework
sqlite3
drf-yasg
